+++ lib/risu/templates/pci_compliance.rb
@@ -34,7 +34,7 @@ module Risu
 				{
 					:name => "pci_compliance",
 					:author => "hammackj",
-					:version => "0.0.4",
+					:version => "0.0.5",
 					:description => "Generates a PCI Compliance Overview Report"
 				}
 			end
@@ -49,9 +49,12 @@ module Risu
 
 				text "\n\n\n"
 
-				@hosts_count = Host.find(:all, :conditions => ["pci_dss_compliance is not null"]).count
-				@hosts_passed = Host.find(:all, :conditions => ["pci_dss_compliance like 'passed'"])
-				@hosts_failed = Host.find(:all, :conditions => ["pci_dss_compliance like 'failed'"])
+				#@hosts_count = Host.find(:all, :conditions => ["pci_dss_compliance is not null"]).count
+				@hosts_count = HostProperty.where(:name => "pci-dss-compliance").count
+				#@hosts_passed = Host.find(:all, :conditions => ["pci_dss_compliance like 'passed'"])
+				@hosts_passed = HostProperty.where(:name => "pci-dss-compliance").where(:value => "passed")
+				#@hosts_failed = Host.find(:all, :conditions => ["pci_dss_compliance like 'failed'"])
+				@hosts_failed = HostProperty.where(:name => "pci-dss-compliance").where(:value => "failed")
 
 				output.font_size(20) do
 					output.text "Summary\n", :style => :bold
@@ -70,7 +73,8 @@ module Risu
 
 					output.text "\n"
 
-					@hosts_passed.each do |host|
+					@hosts_passed.each do |host_prop|
+						host = host_prop.host
 						text "#{host.ip} / #{host.fqdn} - passed\n"
 					end unless @hosts_passed == nil
 
@@ -86,8 +90,9 @@ module Risu
 
 					text "\n"
 
-					@hosts_failed.each do |host|
-						host_id = host.id
+					@hosts_failed.each do |host_prop|
+						host_id = host_prop.host_id
+						host = host_prop.host
 						plugin = Plugin.find(:first, :conditions => { :id => 33929 })
 						item = Item.find(:first, :conditions => { :host_id => host_id, :plugin_id => plugin.id })